refactor: keep http-related code together
